refactor: add hint controller move to hint logic from game controller
